Clean up more long JSX settings files
