+++ src/renderer/src/components/sidebar/AddRepoDialog.tsx
@@ -24,7 +24,8 @@ import { Tooltip, TooltipContent, TooltipTrigger } from '@/components/ui/tooltip
 import { Input } from '@/components/ui/input'
 import { NestedRepoTreePreview } from '@/components/repo/NestedRepoTreePreview'
 import { track } from '@/lib/telemetry'
-import { RemoteStep, CloneStep, useRemoteRepo } from './AddRepoSteps'
+import { CloneStep, useRemoteRepo } from './AddRepoSteps'
+import { RemoteStep } from './AddRepoRemoteStep'
 import { CreateStep, useCreateRepo } from './AddRepoCreateStep'
 import { getProjectAddedPrimaryBranchName, SetupStep } from './AddRepoSetupStep'
 import { getCloneDestinationAutoFill } from './clone-defaults'
